+++ src/backend/node/commands/CS.java
@@ -20,8 +20,8 @@ public class CS extends ZeroArgumentNode {
 	public Response run(ManipulateController sharedHandle) {
 		
 		LambdaInterface l = (Turtle t) -> {
-			TurtleTransferObject dto = new TurtleTransferObject(false, t.getID(), false, t.isPenDown(), t.getPosition(), new int[]{0,0});
-			t.setPosition(new int[]{0,0});
+			TurtleTransferObject dto = new TurtleTransferObject(false, t.getID(), false, t.isPenDown(), t.getPosition(), new double[]{0,0});
+			t.setPosition(new double[]{0,0});
 			t.notifyObservers(dto);
 		};
